requirements.txt
streamlit==0.86.0
tensorflow==2.7.0
pillow==8.4.0
numpy==1.21.5
matplotlib